test: add some more unit tests and cleanup array

+++ src/Jc.OpenNov/Utilities/BinaryExtensions.cs
@@ -55,7 +55,7 @@ public static class ByteBufferExtensions
             throw new EndOfStreamException();
 
         // Big-endian: combine 4 bytes
-        return (int)((bytes[0] << 24) | (bytes[1] << 16) | (bytes[2] << 8) | bytes[3]);
+        return (bytes[0] << 24) | (bytes[1] << 16) | (bytes[2] << 8) | bytes[3];
     }
 
     public static void PutInt(this BinaryWriter writer, int value)
